move congruence to arithmetic
